Move project source to src folder
