+++ src/components/index/calendar-view-item-detail.vue
