+++ src/pages/privacy.vue
