+++ src/plugins/google-api/config.js
