+++ src/plugins/google-api/core.js
