+++ src/static/README.md
